+++ assets/DarkStyleAsset.php
@@ -6,6 +6,9 @@ use humhub\modules\ui\view\helpers\ThemeHelper;
 use Yii;
 use yii\web\AssetBundle;
 
+/*
+ * DarkStyleAsset is the default dark style, based on browser/system preference
+ */
 class DarkStyleAsset extends AssetBundle
 {
     public $publishOptions = [
@@ -14,12 +17,14 @@ class DarkStyleAsset extends AssetBundle
 
     public $sourcePath = '@dark-mode/themes/DarkHumHub';
     
+    // Tell browser to use stylesheet only when in dark mode
     public $css = [
         ['css/theme.css', 'media' => 'screen and (prefers-color-scheme: dark)'],
     ];
     
     public function init()
     {
+        // Find theme selected by module settings
         $themeName = Yii::$app->getModule('dark-mode')->settings->get('theme', 'DarkHumHub');
         $theme = ThemeHelper::getThemeByName($themeName);